adjustments to the leaflet grid, new leaflet list format wip
