+++ components/ShareOptions/index.tsx
@@ -6,7 +6,7 @@ import { useSmoker } from "components/Toast";
 import { Menu, MenuItem } from "components/Layout";
 import { ActionButton } from "components/ActionBar/ActionButton";
 import useSWR from "swr";
-import { useTemplateState } from "app/home/CreateNewButton";
+import { useTemplateState } from "app/home/Actions/CreateNewButton";
 import LoginForm from "app/login/LoginForm";
 import { CustomDomainMenu } from "./DomainOptions";
 import { useIdentityData } from "components/IdentityProvider";
